Update C++ standard for tests from C++11 to C++14 (#293)
Since 1.13.0, gtest requires at least C++14. When it is automatically
downloaded and compiled, gtest’s CMake scripts override the C++
standard, so the tests are compiled in C++14 mode already. Changing the
C++ standard in cpu_features will make it easier for users such as Linux
distribution packagers who are linking against an external or
system-wide copy of gtest 1.13.0 or later.